backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-jsoneditor==0.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
flake8==7.1.1
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
reportlab==4.2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
